cleaned up code, removed circular dependencies, moved seeds to own file, handle request errors
