--- src/init.js
@@ -1,25 +1,0 @@
-const Network = require("./network.js");
-const Api = require("./api.js");
-const Peer = require("./peer.js");
-
-const Init = {};
-
-Init.initP;
-
-Init.init = (networkName) => {
-    Init.initP = new Promise((resolve, reject) => {
-        Network.useNet(networkName).then(() => {
-            Peer.getPeersList((err, success, response) => {
-                if(response != null && response.success)
-                {
-                    Network.peers = response.peers.filter((peer) => peer.status == "OK");
-                    resolve();
-                }
-                else
-                    reject("Error getting peers");
-            });
-        }).catch((err) => reject(err));
-    });
-}
-
-module.exports = Init;
